+++ Forge/src/main/java/com/hypherionmc/craterlib/common/ForgeCompatHelper.java
@@ -42,4 +42,14 @@ public class ForgeCompatHelper implements CompatUtils {
         IBleeding bleeding = PlayerReviveServer.getBleeding(player.toMojangServerPlayer());
         return bleeding != null && bleeding.revived();
     }
+
+    @Override
+    public boolean isPrivateMessage(BridgedPlayer player) {
+        return false;
+    }
+
+    @Override
+    public Component getChannelPrefix(BridgedPlayer player) {
+        return Component.empty();
+    }
 }